text2text/requirements.txt
fastapi==0.115.5
uvicorn==0.20.0
llama_cpp_python==0.3.2